:lipstick:
Tackle all the clippy nits
