+++ crates/repl/src/notebook.rs
@@ -1,7 +1,5 @@
 mod cell;
-mod notebook;
 mod notebook_ui;
 pub(crate) mod static_sample;
 pub use cell::*;
-pub use notebook::*;
 pub use notebook_ui::*;
