--- crates/repl/src/notebook/design_decisions.md
@@ -1,3 +1,0 @@
-This document will solidify the design decisions made for the project in a central place for reference, and can be deleted before the PR is finished.
-
-* Support the most recent notebook format by major version, show an error if not that. Users can use a separate tool to migrate super old notebooks (for now).
